salvage: address /challenge findings on 861de0d + 1a9424c
Round-1 adversarial review surfaced three real defects in the recently
landed durability + trainer simplify commits.

Finding 3 (duplicates samples on crash mid-cleanup) + Finding 5 (worker
retry after crash double-counts):

The pre-existing cleanup-after-flush ordering had a window where, between
the compacted shard being written and the contributing pending zarrs
being deleted, a crash would leave both states on disk. Recovery had no
way to distinguish "crashed before commit" (re-seed) from "crashed after
commit" (delete) and re-seeded both, doubling samples in replay.

Replace post-flush cleanup with an in-flight state machine:

  _pending/<int>_<full-sha>_<token>.zarr
    → _in_flight/<flush_token>/<basename>.zarr (atomic same-fs rename)
    → _compacted/...<flush_token>.zarr (commit point — token in filename)
    → delete _in_flight/<flush_token>/

Recovery now token-matches each in-flight group against _compacted/. If
matched, delete the leftover staging dir. If unmatched, move the contents
back to _pending so phase-2 re-seeds them. Pending filename now embeds
the full upload sha so phase-2 also backfills recent_upload_shas — a
worker retry after the crash is then deduped against the recovered
samples instead of double-counting.

Pending filenames with the same sha are treated as orphaned duplicates
(left over from a silently-failed delete on the duplicate-upload path):
the first wins, the rest are deleted without re-seeding. Closes the
related Finding 4 hole as part of the F5 logic.

Finding 6 (test gaps): three new regressions cover the failure shapes
the original 4 tests omit — crash mid-cleanup with token-matched
compacted shard, worker retry after crash, orphan-duplicate pending.

Finding 1 (autocast dtype divergence): the trainer.py simplify routed
_amp_context through inference_autocast(... dtype="auto"), which falls
back to FP16 on non-BF16 CUDA cards. Training has no GradScaler, so the
fallback would silently underflow gradients. Pin dtype="bf16"
explicitly to preserve the pre-commit invariant.

429 tests pass (422 base + 7 durability), lint clean.

+++ chess_anti_engine/train/trainer.py
@@ -461,7 +461,10 @@ class Trainer:
         )
 
     def _amp_context(self):
-        return inference_autocast(device=self.device, enabled=self.use_amp)
+        # Pinned to bf16: training has no GradScaler, so an FP16 fallback
+        # would silently underflow gradients on non-BF16 CUDA cards. The
+        # ``inference_autocast`` helper would auto-fallback to FP16 there.
+        return inference_autocast(device=self.device, enabled=self.use_amp, dtype="bf16")
 
     @staticmethod
     def _extract_loss_scalars(losses: dict[str, torch.Tensor]) -> dict[str, float]:
